+++ src/main/java/thunder/hack/features/hud/impl/WaterMark.java
@@ -1,8 +1,7 @@
 package thunder.hack.features.hud.impl;
 
-import net.minecraft.client.render.RenderPipelines;
-
 import net.minecraft.client.gl.RenderPipelines;
+//
 import org.joml.Matrix3x2fStack;
 import com.mojang.blaze3d.systems.RenderSystem;
 import net.minecraft.client.gui.DrawContext;
@@ -121,9 +120,9 @@ public class WaterMark extends HudElement {
 
             Render2DEngine.addWindow(context.getMatrices(), getPosX(), getPosY(), getPosX() + 100, getPosY() + 64, 1f);
             context.getMatrices().pushMatrix();
-            context.getMatrices().translate(getPosX() + 10, getPosY() + 32);
+            context.getMatrices().translate((float) (getPosX() + 10), (float) (getPosY() + 32));
             context.getMatrices().rotate((float) Math.toRadians(mc.player.age * 3 + Render3DEngine.getTickDelta(false)));
-            context.getMatrices().translate(-(getPosX() + 10), -(getPosY() + 32));
+            context.getMatrices().translate((float) (-(getPosX() + 10)), (float) (-(getPosY() + 32)));
             context.drawTexture(RenderPipelines.GUI_TEXTURED, TextureStorage.baltika, (int) getPosX() - 10, (int) getPosY() + 2, 0, 0, 40, 64, 40, 64, -1);
             context.getMatrices().popMatrix();
             Render2DEngine.popWindow();
